refactor: RecipeCardのstoriesのリファクタリング
