+++ frontend/src/lib/types/recipe.ts
@@ -1,0 +1,12 @@
+export type Recipe = {
+  name: string;
+  imageUrl: string;
+  recipe: {
+    ingredients: Array<{
+      name: string;
+      amount: string;
+    }>;
+    steps: string[];
+    tips: string[];
+  };
+}
