+++ frontend/src/pages-layout/Recipes/Recipes.stories.ts
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
 import { SPStory } from "@/lib/storybook/SPStory";
-import { recipeCardArgs } from "@/lib/storybook/recipeCardArgs";
+import { recipeCardArgs } from "@/lib/storybook/recipeArgs";
 import { Recipes } from "@/pages-layout/Recipes/Recipes";
 
 const meta: Meta<typeof Recipes> = {
